interest/dcf conversion

+++ src/main/java/org/blacksmith/finlib/curve/types/Knot.java
@@ -4,11 +4,11 @@ import lombok.Value;
 
 @Value(staticConstructor = "of")
 public class Knot implements Comparable<Knot>{
-  int x;
+  double x;
   double y;
 
   @Override
   public int compareTo(Knot o) {
-    return Integer.compare(this.x, o.getX());
+    return Double.compare(this.x, o.getX());
   }
 }
